feat: gui improvements

+++ src/Sberbank2Excel/sberbankPDF2Excel.py
@@ -17,7 +17,7 @@ from Sberbank2Excel import version_info
 def sberbankPDF2Excel(input_file_name: str,
                       output_file_name: Union[str, None] = None,
                       format: str = 'auto',
-                      pdf_lib: str = 'pdfminer',
+                      pdf_lib: str = 'auto',
                       leave_intermediate_txt_file: bool = False,
                       perform_balance_check=True,
                       output_file_type: str = "xlsx",
@@ -41,6 +41,8 @@ def sberbankPDF2Excel(input_file_name: str,
     """
 
     print(f"{format=}")
+    if pdf_lib == "auto":
+        pdf_lib = None
 
     print("*"*30)
     print("Конвертируем файл " + input_file_name)
@@ -63,22 +65,32 @@ def sberbankPDF2Excel(input_file_name: str,
         output_file_name = path 
 
     try:
-        if extension == ".pdf":
-            if pdf_lib == "pdfminer":
-                pdf_2_txt_file_pdfminer(input_file_name, tmp_txt_file_name)
-            elif pdf_lib == "pypdf":
-                pdf_2_txt_file_pypdf(input_file_name, tmp_txt_file_name)
+        has_true = False
+        errors = []
+        for pdf_lib_c in ["pdfminer", "pypdf"] if pdf_lib is None else [pdf_lib]:
+            try:
+                if extension == ".pdf":
+                    if pdf_lib_c == "pdfminer":
+                        pdf_2_txt_file_pdfminer(input_file_name, tmp_txt_file_name)
+                    elif pdf_lib_c == "pypdf":
+                        pdf_2_txt_file_pypdf(input_file_name, tmp_txt_file_name)
 
-        created_file_name = sberbankPDFtext2Excel(tmp_txt_file_name,
-                                                  output_file_name,
-                                                  pdf_lib=pdf_lib,
-                                                  format=format,
-                                                  perform_balance_check = perform_balance_check,
-                                                  output_file_type=output_file_type,
-                                                  reversed_transaction_order=reversed_transaction_order)
-
-        if (not leave_intermediate_txt_file) and (not extension == ".txt"):
-            os.remove(tmp_txt_file_name)
+                created_file_name = sberbankPDFtext2Excel(tmp_txt_file_name,
+                                                        output_file_name,
+                                                        pdf_lib=pdf_lib_c,
+                                                        format=format,
+                                                        perform_balance_check = perform_balance_check,
+                                                        output_file_type=output_file_type,
+                                                        reversed_transaction_order=reversed_transaction_order)
+                has_true = True
+            except Exception as e:
+                errors.append(e)
+                
+        if has_true:
+            if (not leave_intermediate_txt_file) and (not extension == ".txt"):
+                os.remove(tmp_txt_file_name)
+        else:
+            raise Exception(*errors)
 
     except:
         raise
